+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix the tree structure generation of the `al-toc` element
 * Make `TableOfContents#entries` return a flat list and `TableOfContents#tree` a tree structure
+* Also take siblings of heading elements into account for the `TableOfContents` scrollspy
 
 ## 0.2.3 (2023-01-23)